Add How It Works CRUD and price of service

+++ app/Http/Controllers/HowItWorksController.php
@@ -3,37 +3,35 @@
 namespace App\Http\Controllers;
 
 use App\Models\Attachment;
-use App\Models\Service;
+use App\Models\HowItWorks;
 use Illuminate\Http\Request;
 use Illuminate\Http\Response;
 use Illuminate\Support\Facades\DB;
-use Illuminate\Support\Str;
 use Ramsey\Uuid\Uuid;
 use Yajra\DataTables\Facades\DataTables;
 
-class BackendServiceController extends Controller
+class HowItWorksController extends Controller
 {
     /**
      * Display a listing of the resource.
      */
     public function index()
     {
-        return view('backend.home_page.services.index');
+        return view('backend.home_page.how_it_works.index');
     }
 
     public function dataTable()
     {
-        $query = Service::query();
+        $query = HowItWorks::query();
         return DataTables::eloquent($query)
             ->addIndexColumn()
-            ->addColumn('action', function(Service $service) {
-                $btn ='<a href="'.route('services.edit',['service'=>$service->id]).'" class="btn btn-primary bg-gradient-primary btn-sm btn-edit"><i class="fa fa-edit"></i></a>
-                        <a role="button" data-id="'.$service->id.'" class="btn btn-danger btn-sm btn-delete"><i class="fa fa-trash"></i></a>';
+            ->addColumn('action', function(HowItWorks $howItWorks) {
+                $btn ='<a href="'.route('how-it-works.edit',['how_it_work'=>$howItWorks->id]).'" class="btn btn-primary bg-gradient-primary btn-sm btn-edit"><i class="fa fa-edit"></i></a>
+                        <a role="button" data-id="'.$howItWorks->id.'" class="btn btn-danger btn-sm btn-delete"><i class="fa fa-trash"></i></a>';
                 return $btn;
             })
-            ->addColumn('image', function(Service $service) {
-                return '<img height="100px" src="'.asset($service->attachments->first()->file ?? '').'">';
-
+            ->addColumn('image', function(HowItWorks $howItWorks) {
+                return '<img height="100px" src="'.asset($howItWorks->attachments->file ?? '').'">';
             })
             ->rawColumns(['action','image'])
             ->toJson();
@@ -44,23 +42,20 @@ class BackendServiceController extends Controller
      */
     public function create()
     {
-        return view('backend.home_page.services.create');
+        return view('backend.home_page.how_it_works.create');
     }
-
     /**
      * Store a newly created resource in storage.
      */
     public function store(Request $request)
     {
-
         // Validate the request data
         $validatedData = $request->validate([
-            'title' =>[
-                'required','max:255',
+            'name' => [
+                'required', 'max:255',
             ],
-            'description' => 'nullable',
+            'details' => 'nullable',
             'status' => 'required|boolean',
-            'price' => 'nullable',
             'attachments' => 'nullable',
         ]);
 
@@ -70,22 +65,20 @@ class BackendServiceController extends Controller
         try {
 
             // Create a new Product record in the database
-            $service = Service::create([
-                'title' => $validatedData['title'],
-                'description' => $validatedData['description'],
-                'price' => $validatedData['price'],
+            $howItWorks = HowItWorks::create([
+                'name' => $validatedData['name'],
+                'details' => $validatedData['details'],
                 'status' => $validatedData['status'],
-                'slug' => Str::slug($validatedData['title'], '-'),
             ]);
 
             if ($request->file('attachments')) {
                 $counter = 0;
                 foreach ($request->file('attachments') as $key => $attachmentFile) {
-                    $filename = Uuid::uuid1()->toString().$service->id.'-'.$key.'.' .$attachmentFile->extension();
+                    $filename = Uuid::uuid1()->toString() . $howItWorks->id . '-' . $key . '.' . $attachmentFile->extension();
 
-                    $destinationPath = 'uploads/attachments/services';
+                    $destinationPath = 'uploads/attachments/how_it_works';
                     $attachmentFile->move(public_path($destinationPath), $filename);
-                    $path = 'uploads/attachments/services/' . $filename;
+                    $path = 'uploads/attachments/how_it_works/' . $filename;
 
 
                     $attachment = new Attachment([
@@ -93,7 +86,7 @@ class BackendServiceController extends Controller
                         'file' => $path,
                         'sort' => $request->attachment_sort[$counter],
                     ]);
-                    $service->attachments()->save($attachment);
+                    $howItWorks->attachments()->save($attachment);
 
 
                     $counter++;
@@ -106,9 +99,9 @@ class BackendServiceController extends Controller
 
             // Redirect to the index page with a success message
             return response()->json([
-                'status'=>true,
-                'redirect_url'=>route('services.index'),
-                'message'=>'Service created successfully',
+                'status' => true,
+                'redirect_url' => route('how-it-works.index'),
+                'message' => 'Data Saved successfully',
             ]);
         } catch (\Exception $e) {
             // Roll back the transaction in case of an error
@@ -116,38 +109,37 @@ class BackendServiceController extends Controller
 
             // Handle the error and redirect with an error message
             return response()->json([
-                'status'=>false,
-                'message'=>$e->getMessage(),
+                'status' => false,
+                'message' => $e->getMessage(),
             ]);
         }
     }
 
-
     /**
      * Show the form for editing the specified resource.
      */
-    public function edit(Service $service)
+    public function edit(HowItWorks $howItWork)
     {
         try {
-            return view('backend.home_page.services.edit', compact('service',));
+            return view('backend.home_page.how_it_works.edit', compact('howItWork',));
         } catch (\Illuminate\Database\Eloquent\ModelNotFoundException $e) {
             // Handle the case where the Product is not found
-            return redirect()->route('services.index')->with('error',$e->getMessage());
+            return redirect()->route('how-it-works.index')->with('error',$e->getMessage());
         }
+
     }
 
     /**
      * Update the specified resource in storage.
      */
-    public function update(Request $request, Service $service)
+    public function update(Request $request, HowItWorks $howItWork)
     {
         $validatedData = $request->validate([
-            'title' =>[
-                'required','max:255',
+            'name' => [
+                'required', 'max:255',
             ],
-            'description' => 'nullable',
-            'price' => 'nullable',
-            'status' => 'required|boolean', // Ensure 'status' is boolean
+            'details' => 'nullable',
+            'status' => 'required|boolean',
         ]);
 
         // Start a database transaction
@@ -155,23 +147,22 @@ class BackendServiceController extends Controller
 
         try {
 
-            $service->update([
-                'title' => $validatedData['title'],
-                'description' => $validatedData['description'],
-                'price' => $validatedData['price'],
+            $howItWork->update([
+                'name' => $validatedData['name'],
+                'details' => $validatedData['details'],
                 'status' => $validatedData['status'],
-                'slug' => Str::slug($validatedData['title'], '-'),
             ]);
             if ($request->file('attachments')) {
                 $counter = 0;
                 foreach ($request->file('attachments') as $key => $attachmentFile) {
-                    $filename = Uuid::uuid1()->toString().$service->id.'-'.$key.'.' .$attachmentFile->extension();
+                    $filename = Uuid::uuid1()->toString().$howItWork->id.'-'.$key.'.' .$attachmentFile->extension();
 
-                    $destinationPath = 'uploads/attachments/services';
+                    $destinationPath = 'uploads/attachments/how_it_works';
                     $attachmentFile->move(public_path($destinationPath), $filename);
-                    $path = 'uploads/attachments/services/' . $filename;
+                    $path = 'uploads/attachments/how_it_works/' . $filename;
 
-                    $service->attachments()->create([
+
+                    $howItWork->attachments()->create([
                         'user_id' => auth()->id(),
                         'file' => $path,
                         'sort' => $request->attachment_sort[$counter],
@@ -187,8 +178,8 @@ class BackendServiceController extends Controller
             // Redirect to the index page with a success message
             return response()->json([
                 'status'=>true,
-                'redirect_url'=>route('services.index'),
-                'message'=>'Service Updated successfully',
+                'redirect_url'=>route('how-it-works.index'),
+                'message'=>'Data Updated successfully',
             ]);
         } catch (\Exception $e) {
             // Roll back the transaction in case of an error
@@ -202,27 +193,24 @@ class BackendServiceController extends Controller
         }
     }
 
+
     /**
      * Remove the specified resource from storage.
      */
-    public function destroy(Service $service)
+    public function destroy(HowItWorks $howItWork)
     {
         try {
             // Delete the Product record
-            foreach ($service->attachments as $attachment){
+            if ( $attachment = $howItWork->attachments){
                 if (file_exists(public_path($attachment->file))){
                     unlink(public_path($attachment->file));
                 }
                 $attachment->delete();
             }
 
-            if (file_exists(public_path($service->icon))){
-                unlink(public_path($service->icon));
-            }
-
-            $service->delete();
+            $howItWork->delete();
             // Return a JSON success response
-            return response()->json(['success'=>true,'message' => 'Service deleted successfully'], Response::HTTP_OK);
+            return response()->json(['success'=>true,'message' => 'Data deleted successfully'], Response::HTTP_OK);
         } catch (\Exception $e) {
             // Handle any errors, such as record not found
             return response()->json(['success'=>false,'message' =>$e->getMessage()], Response::HTTP_NOT_FOUND);
